Create NMW Policy Impact Based on Region within UK
The imposition of mandatory minimum wage is a common instrument of economic policy (Fidrmuc and Tena, 2018). Since its inception on 1 April 1999 (Lourie, 1999), the UK minimum wage has been implemented nationally, applying to all parts of the country (Dolton, Rosazza-Bondibene and Wadsworth, 2009).

The central idea of this report is to assess this regional variation National Minimum Wage (NMW) policy’s impact on hourly pay throughout the UK. The research question is therefore: Did The Introduction Of the National Minimum Wage Policy Affect The
Hourly Pay Based On Region, Specifically London and North West?