Rename NMW Policy Impact on Region to NMW Policy Impact Based on Region
The central idea of this code file is to assess this regional variation National Minimum Wage (NMW) policy’s impact on hourly pay throughout the UK. The research question is therefore: Did The Introduction Of the National Minimum Wage Policy Affect The Hourly Pay Based On Region, Specifically London and North West?

To answer this research question, we conducted ordinary least squares regression and matching to try and estimate the causal effect of working in London on hourly wages as compared to being in the North West. The results show that the hourly wage in London is indeed higher, although the magnitude of such an effect differs between both methods.